jwt 처리중 9

+++ src/main/java/kr/co/hanipactor/application/store/StoreMapper.java
@@ -5,5 +5,5 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface StoreMapper {
     // int save(StorePostReq req); // 가게등록
-    Integer findStoreIdByUserId(int userId);
+    Integer findStoreIdByUserId(long userId);
 }
